requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-bootstrap-themes==3.1.2
django-bootstrap3==6.2.2
django-debug-toolbar==1.4
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6.1
sqlparse==0.1.16
static3==0.6.1
wheel==0.24.0
whitenoise==2.0.4
djangorestframework==3.2.4
django-rest-swagger==0.3.4
PyYAML==3.11
requests==2.8.1
django-cors-headers==1.1.0
